.github/workflows, images/ubuntu: Add p11-kit-client.so to newer Ubuntus
A subsequent commit will use this to give Toolbx containers access to
the certificates from certificate authorities on the host.  This commit
was kept separate from the changes to toolbox(1) to ensure that the
ubuntu-toolbox images are ready before that happens.

Unfortunately, Ubuntu 16.04 Xenial Xerus has p11-kit 0.23.2 [1], while
p11-kit-client.so was introduced in p11-kit 0.23.10 [2].  Therefore,
this feature will be disabled for Ubuntu 16.04.

This reverts parts of commit 1e9308625a70eb79503c3c1e7cf2b19c516261f4 to
re-enable building the ubuntu-toolbox:18.04 image.

[1] https://code.launchpad.net/~git-ubuntu-import/ubuntu/+source/p11-kit/+git/p11-kit/+ref/ubuntu/xenial-updates

[2] p11-kit commit 0684cd7b7f815b41
    https://github.com/p11-glue/p11-kit/commit/0684cd7b7f815b41
    https://github.com/p11-glue/p11-kit/pull/15

https://github.com/containers/toolbox/issues/626