Isolate the host's XDG_RUNTIME_DIR from the system tests
XDG_RUNTIME_DIR is needed for two groups of reasons when Toolbx is used
rootless.

First, it's important for toolbox(1) itself to work rootless because it
needs to place several files:

  * The 'lock' file to synchronize Podman migrations.

  * The initialization stamp file to synchronize the container's entry
    point with the user-facing 'enter' and 'run' commands running on the
    host operating system.

  * The generated Container Device Interface specification.

These files need to be separate for the toolbox(1) processes run by the
system tests, those run by the user for 'normal' use, and concurrent
invocations of the tests.

Therefore, it's better to use a custom XDG_RUNTIME_DIR that's within the
sandbox offered by Bats [1].  The sandbox is clearly labelled as being
used by Bats, is unique for each invocation, and Bats takes care of
cleaning everything up once it has finished running.

Note that XDG_RUNTIME_DIR's Unix access mode MUST be 0700 [2].  eg.,
Ubuntu 22.04 and 24.04 Desktop have a umask of 0002, and if an access
mode is not explicitly specified, XDG_RUNTIME_DIR will be created with
0775.  That will cause dbus-daemon(1) to fail with:
  Unable to set up transient service directory: XDG_RUNTIME_DIR
      "/var/tmp/bats-run-4XQL6i/suite/xdg-runtime-dir" can be written by
      others (mode 040775)

Second, XDG_RUNTIME_DIR is used to propagate things like the user D-Bus,
Pipewire and Wayland sockets from the host to the container.  These
don't need to be separated.  However, if a custom XDG_RUNTIME_DIR is
used then those sockets that are used by the system tests, such as the
user D-Bus socket, have to be replicated.

Therefore, a custom D-Bus instance is run to offer the user D-Bus socket
with a configuration similar to that of the host OS.  The dbus-daemon(1)
implementation is used for the sake of simplicity.  It creates the
socket itself based on the configuration, unlike dbus-broker-launch(1)
where the socket must be separately created and passed to it by its
parent.

However, Podman can't use systemd as the cgroups manager with this D-Bus
instance, as the bus wasn't started by the user systemd instance.  So, a
custom containers.conf(5) is used to change the cgroups manager to
cgroupfs.  The only other options in the containers.conf(5) are those
that are common across Fedora 41 and 42, and Ubuntu 22.04 and 24.04.

[1] https://bats-core.readthedocs.io/en/stable/writing-tests.html

[2] https://specifications.freedesktop.org/basedir-spec/latest/

https://github.com/containers/toolbox/pull/1652