pkg/utils, test/system: Isolate the host's HOME from the system tests
Currently, the test for the 'run' command to ensure that a login shell
is used to invoke commands temporarily changes the user's
~/.bash_profile [1].  This is really inappropriate because the system
tests shouldn't touch the user's configuration or data.

Therefore, it's better to use a custom home directory for the system
tests that's within the sandbox offered by Bats [2].

Some of the common base directories that derive from the user's home
directory [3] were redefined in terms of the custom home directory.
Podman [4,5,6], Toolbx [7] and many other programs rely on these for
various purposes, and Bats might get invoked with different values set
in the environment.  The other option is to unset the variables from the
environment.

This was made possible by commit 3bf9fdde4ce12321 that unbroke the
overriding of the HOME environment variable in the Go implementation.
Otherwise, the system tests failed with:
  not ok 8 help: Try unknown command (forwarded to host)
  # (from function `assert_line' in file
      test/system/libs/bats-assert/src/assert.bash, line 488,
  #  in test file test/system/002-help.bats, line 136)
  #   `assert_line --index 0 "Error: unknown command \"foo\" for
      \"toolbox\""' failed
  #
  # -- line differs --
  # index    : 0
  # expected : Error: unknown command "foo" for "toolbox"
  # actual   : Error: crun: chdir to
      `/var/tmp/bats-run-5pDdU0/suite/home`: No such file or
      directory: OCI runtime attempted to invoke a command that was
      not found
  # --
  #

[1] Commit 1ce59a6a2d3a4c50
    https://github.com/containers/toolbox/commit/1ce59a6a2d3a4c50
    https://github.com/containers/toolbox/pull/1148
    https://github.com/containers/toolbox/issues/1076

[2] https://bats-core.readthedocs.io/en/stable/writing-tests.html

[3] https://specifications.freedesktop.org/basedir-spec/latest/

[4] https://docs.podman.io/en/latest/markdown/podman.1.html

[5] https://github.com/containers/common/blob/main/docs/containers.conf.5.md

[6] https://github.com/containers/storage/blob/main/docs/containers-storage.conf.5.md

[7] https://github.com/containers/toolbox/blob/main/doc/toolbox.conf.5.md

https://github.com/containers/toolbox/pull/1668