test/system: Restore some configuration paths to their usual values
The default location for a rootless user's storage.conf file [1], which
is overridden by the CONTAINERS_STORAGE_CONF environment variable [2],
is at $XDG_CONFIG_HOME/containers/storage.conf, and the default value of
the rootless_storage_path option is $XDG_DATA_HOME/containers/storage.
Earlier they were being set to different paths because the user's home
directory on the host operating system was not isolated from the system
tests.

Now that the user's home directory is isolated because the system tests
use a custom HOME, these configuration paths can be restored to their
default values relative to this custom HOME.  This will make things more
understandable by making the directory hierarchy within Bats' sandbox
similar to the defaults.

The CONTAINERS_STORAGE_CONF environment variable was not removed to
protect against cases where Bats might get invoked with a different
CONTAINERS_STORAGE_CONF set in the environment.  The other option is to
unset the variable from the environment.  The former was chosen to
deviate as little from the status quo as possible.

[1] https://github.com/containers/storage/blob/main/docs/containers-storage.conf.5.md

[2] https://docs.podman.io/en/latest/markdown/podman.1.html

https://github.com/containers/toolbox/pull/1673