cmd/initContainer: Fail if non-folder files can't be removed for linking
There's no reason to ignore an error when trying to remove a file within
the container that's not a directory, before turning it into a symbolic
link.

The POSIX shell implementation didn't make any distinction between
directories and other types of files.

For files that aren't directories, it did:
  cd /path/to \
  && rm --force file \
  && ln --symbolic /run/host/path/to/file file

For directories, it did:
  rmdir /path/to/directory \
  && mkdir --parents /path/to/target \
  && ln --symbolic /path/to/target /path/to/directory

It's possible that this was a misunderstanding about the behaviour of
'rm --force' when writing the Go implementation.  It only ignores errors
arising from missing files, and not every error [1].  eg., if the file
is a mount point, it won't ignore the error:
  $ sudo mount --rbind /etc/machine-id foo
  $ rm --force foo
  rm: cannot remove 'foo': Device or resource busy

Fallout from 772b66bf3e962785730e55f6242fa74066de0743

[1] https://man7.org/linux/man-pages/man1/rm.1.html

https://github.com/containers/toolbox/pull/1692