cmd, test/system: Deprecate containers using o.fd.Flatpak.SessionHelper
Starting from Toolbx 0.0.97, released in November 2020, Toolbx
containers no longer need the org.freedesktop.Flatpak.SessionHelper
D-Bus interface.  This change was made to enable running toolbox(1) as
root [1].

It is going to be five years in a few months, since this change was
made, and it's becoming harder and harder to practically test backwards
compatibility with these containers.  All the host operating systems
where Toolbx is regularly tested offer newer versions:
  * Arch Linux and Fedora offer the latest upstream releases
  * RHEL 8.10 offers at least 0.0.99.4
  * Ubuntu offers at least 0.0.99.2 [2]

Therefore, this is a good time to mark these older containers as
deprecated, as has been done in the past in similar cases [3].  A
deprecation notice is the precursor to actually dropping support for
these old containers in the future.

[1] Commit 71b5c8c0a235249b
    https://github.com/containers/toolbox/commit/71b5c8c0a235249b
    https://github.com/containers/toolbox/pull/591
    https://github.com/containers/toolbox/issues/267

[2] https://packages.ubuntu.com/jammy/podman-toolbox

[3] Commit 9dc52814301c5a3f
    https://github.com/containers/toolbox/commit/9dc52814301c5a3f
    https://github.com/containers/toolbox/pull/336

https://github.com/containers/toolbox/issues/1684