test/system: Tighten the regular expression used to check the version
The MAJOR version will always be 0, the MINOR version can't be 0 after
the release of 0.1.0; until 1.0.0 or 1.0 is released, which won't happen
in the short-term future.  Similarly, the MICRO version can't be 0 after
the release of 0.1.1, until 0.2.0 is released.

Future releases will default to not having a MICRO version and use a
MAJOR.MINOR versioning scheme.  A MICRO version will be reserved for the
same purposes that a NANO version was reserved for until now, and it
will never be 0.

Tighten the regular expression used to check the version to match this
present reality.  It can be revisited when 1.0 is eventually released.

https://github.com/containers/toolbox/pull/1703