pkg/podman/podman, cmd/rmi: Change return values of IsToolboxImage()
The original version of the IsToolboxImage() function in
pkg/podman/podman.go returned an error when an image was not
compatible with Toolbx.

The new version returns 'false' without error on a non-Toolbx
container, so it can be distinguished when the image inspection
actually fails.

This new behavior is used in detecting such images when creating
or running Toolbx containers: f06a819