cmd/initContainer: Unbreak access to CA certificates in su(1) & sudo(8)
When a Toolbx container is set up to use the p11-kit-client.so PKCS #11
module instead of the usual p11-kit-trust.so module, the
P11_KIT_SERVER_ADDRESS environment variable must be set inside the
container, so that it can communicate with the host operating system.

Currently, this works as described above with the 'enter' and 'run'
commands, but not within child sessions started by su(1) [1] and
sudo(8) [2], because P11_KIT_SERVER_ADDRESS is absent.

As a result, when 'trust extract' [3] or update-ca-trust(8) [4] is
invoked as root, several files within /etc/ca-certificates/extracted in
Arch Linux containers or /etc/pki/ca-trust/extracted in Fedora
containers are either emptied out or truncated.  This can happen when
the package manager inside the container notices an update to the
certificates from certificate authorities (or CAs) or the user wants to
add a custom certificate inside the container.

The empty and truncated files in /etc/ca-certificates/extracted or
/etc/pki/ca-trust/extracted break OpenSSL and other cryptographic
components that don't use PKCS #11 to access CA certificates.

To prevent this, P11_KIT_SERVER_ADDRESS must be set in child sessions
started by su(1) and sudo(8).

Fallout from 5ed2442214f0eb8d58d1bfc7df50f62328815cb2

[1] https://www.man7.org/linux/man-pages/man1/su.1.html

[2] https://man7.org/linux/man-pages/man8/sudo.8.html

[3] https://p11-glue.github.io/p11-glue/p11-kit/manual/trust.html

[4] https://www.systutorials.com/docs/linux/man/8-update-ca-trust/

https://github.com/containers/toolbox/issues/626
https://github.com/containers/toolbox/issues/1661