images/arch: Rename a file
A subsequent commit will generate a /etc/sudoers.d configuration snippet
at runtime to set the P11_KIT_SERVER_ADDRESS environment variable.
Adding a suffix to the name of the file clarifies its purpose, and a
numerical prefix clarifies the lexical order in which the files are read
regardless of the rest of their names.

https://github.com/containers/toolbox/issues/626
https://github.com/containers/toolbox/issues/1661