test/system: Check if 'create' created the container
Checking the exit code, and the standard error and output streams of the
'create' command don't say anything about whether the container was
actually created.  Therefore, 'podman ps --all' is used to verify that.

Other tests already do this, so this change makes these tests consistent
with the rest.

Fallout from the following:
  * b27795a03ec8ae3139e2dad93b4efca13e96fb01
  * 983beb13527a9beee1a1b3b2f0022201967c9964
  * ecd1ced719742a6faea4d0cf41d68391f5644b4e

https://github.com/containers/toolbox/pull/1715