pkg/utils: Split the PathExists tests into separate functions
A subsequent commit will ensure that the test functions are stopped if
they fail their set-up.  This means that tests that don't share the same
set-up shouldn't be in the same test function.  Otherwise, a failure in
one test's set-up will prevent the remaining tests that don't need the
failed set-up from running.

https://github.com/containers/toolbox/pull/1727